feat(viz): implement time picker on filter box (#13377)
* refactor(explore): restructure DateFilterControl

* fix UT

* remove datasource props

* remove unused files

* fix DateFilterControl in filterbox over length
