+++ superset-frontend/src/explore/components/controls/DateFilterControl/components/CustomFrame.tsx
@@ -31,17 +31,15 @@ import {
   UNTIL_MODE_OPTIONS,
   MOMENT_FORMAT,
   MIDNIGHT,
-} from '../constants';
-import {
   customTimeRangeDecode,
   customTimeRangeEncode,
   dttmToMoment,
-} from '../utils';
+} from 'src/explore/components/controls/DateFilterControl/utils';
 import {
   CustomRangeKey,
   SelectOptionType,
   FrameComponentProps,
-} from '../types';
+} from 'src/explore/components/controls/DateFilterControl/types';
 
 export function CustomFrame(props: FrameComponentProps) {
   const { customRange, matchedFlag } = customTimeRangeDecode(props.value);
